spliting up main file
-at first, only seperate the serial functions
